requirements.txt
numpy
scipy
pandas
matplotlib
open3d==0.12.0
catkin_pkg
tensorboard
tqdm
trimesh
urdfpy
cython
scikit-image
pykdtree

mmcv-full
mmdet
